+++ Flow/src/utopia/flow/view/immutable/eventful/ChangeFuture.scala
@@ -9,12 +9,10 @@ import utopia.flow.view.mutable.async.VolatileOption
 import utopia.flow.view.template.eventful.Changing
 
 import scala.concurrent.{ExecutionContext, Future}
-import scala.util.{Failure, Success, Try}
+import scala.util.Try
 
 object ChangeFuture
 {
-	// TODO: There should be some easy way to access the optimized constructors without repeating a lot of code
-	
 	/**
 	  * Creates a new future with change events
 	  * @param placeHolder A placeholder value stored until this future completes
@@ -31,18 +29,6 @@ object ChangeFuture
 		new ChangeFuture[A, F](placeHolder, future)(merge)
 	
 	/**
-	  * Creates a new change future
-	  * @param placeHolder A placeholder value returned until the future resolves (successfully)
-	  * @param future A future
-	  * @param exc Implicit execution context
-	  * @tparam A Type of held value
-	  * @return A new change future.
-	  *         This future will continue to contain the placeholder value if the future fails to resolve.
-	  */
-	def apply[A](placeHolder: A, future: Future[A])(implicit exc: ExecutionContext) =
-		merging[A, A](placeHolder, future) { (p, result) => result.getOrElse(p) }
-	
-	/**
 	  * Creates a new change future. Failed future resolves are logged.
 	  * @param placeHolder A placeholder value returned until the future resolves (successfully)
 	  * @param future      A future
@@ -52,7 +38,7 @@ object ChangeFuture
 	  * @return A new change future.
 	  *         This future will continue to contain the placeholder value if the future fails to resolve.
 	  */
-	def caching[A](placeHolder: A, future: Future[A])(implicit exc: ExecutionContext, log: Logger) =
+	def apply[A](placeHolder: A, future: Future[A])(implicit exc: ExecutionContext, log: Logger) =
 		merging[A, A](placeHolder, future) { (p, result) =>
 			result.getOrMap { error =>
 				log(error)
@@ -68,10 +54,11 @@ object ChangeFuture
 	  * @tparam A Type of future content
 	  * @return A new change future (or wrapped future result)
 	  */
+	@deprecated("Please use Changing.wrapFuture(...) instead", "v2.0")
 	def wrap[A](future: Future[A], placeHolder: => A)(implicit exc: ExecutionContext) =
 		future.current match {
 			case Some(v) => Fixed(v)
-			case None => apply[A](placeHolder, future)
+			case None => merging[A, A](placeHolder, future) { (p, result) => result.getOrElse(p) }
 		}
 	
 	/**
@@ -81,28 +68,9 @@ object ChangeFuture
 	  * @tparam A Type of item this future will eventually contain
 	  * @return A new change future
 	  */
+	@deprecated("Please use Changing.wrapFutureToOption(...) instead", "v2.0")
 	def noneUntilCompleted[A](future: Future[A])(implicit exc: ExecutionContext) =
 		wrap(future.map { Some(_) }, None)
-	
-	/**
-	  * Creates a change future that will contain Left until successfully resolved and Right once (if) so.
-	  * Logs possible errors.
-	  * @param placeHolder A placeholder value to return until the future successfully resolves.
-	  * @param future A future
-	  * @param exc Implicit execution context
-	  * @param log A logging implementation
-	  * @tparam A Type of future result & placeholder type
-	  * @return A new change future
-	  */
-	def eitherCaching[A](placeHolder: A, future: Future[A])(implicit exc: ExecutionContext, log: Logger) =
-		merging[Either[A, A], A](Left(placeHolder), future) { (p, result) =>
-			result match {
-				case Success(v) => Right(v)
-				case Failure(error) =>
-					log(error)
-					p
-			}
-		}
 }
 
 /**
@@ -184,17 +152,5 @@ class ChangeFuture[A, F](placeHolder: A, val future: Future[F])(mergeResult: (A,
 			Fixed(f(value))
 		else
 			new ChangeFuture[B, F](f(placeHolder), future)((_, result) => f(mergeResult(value, result)))
-	}
-	override def mergeWith[B, R](other: Changing[B])(f: (A, B) => R) = {
-		if (other.isChanging)
-			MergeMirror.of(this, other)(f)
-		else
-			map { f(_, other.value) }
-	}
-	override def lazyMergeWith[B, R](other: Changing[B])(f: (A, B) => R) = {
-		if (other.isChanging)
-			LazyMergeMirror.of(this, other)(f)
-		else
-			lazyMap { f(_, other.value) }
 	}
 }